requirements.txt
langchain>=0.1.0
langchain-community>=0.0.10
langchain-mistralai>=0.0.1
langchain-google-genai>=0.0.3
faiss-cpu>=1.7.4
pypdf>=3.15.1
python-dotenv>=1.0.0
streamlit>=1.27.0
tqdm>=4.66.1